+++ src/exit.js
@@ -1,23 +1,14 @@
-// Do not destructure so tests can stub it
-import process from 'process'
+import process, { version } from 'process'
 
-// Exit process according to `opts.exitOn` (default: ['uncaughtException']):
-//  - `uncaughtException`: default behavior of Node.js and recommended by
+import semver from 'semver'
+
+// Exit process on `uncaughtException` and `unhandledRejection`
+//  - This is the default behavior of Node.js
+//  - This is  recommended by
 //     https://nodejs.org/api/process.html#process_warning_using_uncaughtexception_correctly
-//  - `unhandledRejection`: default behavior of Node.js since 15.0.0
-// `process.exit()` unfortunately aborts any current async operations and
-// streams are not flushed (including stdout/stderr):
-//  - https://github.com/nodejs/node/issues/784
-//  - https://github.com/nodejs/node/issues/6456
-// We go around this problem by:
-//  - await promise returned by `opts.log()`
-//  - waiting for few seconds (EXIT_TIMEOUT)
-// This last one is a hack. We should instead allow `opts.log()` to return a
-// stream, and keep track of all unique returned streams. On exit, we should
-// then close them and wait for them to flush. We should then always wait for
-// process.stdout|stderr as well.
-export const exitProcess = function (name, exitOn) {
-  if (!exitOn.includes(name)) {
+// This can be disabled with the `keep: true` option.
+export const exitProcess = function (keep, reason) {
+  if (!shouldExit(keep, reason)) {
     return
   }
 
@@ -25,37 +16,27 @@ export const exitProcess = function (name, exitOn) {
   // https://github.com/sinonjs/fake-timers/issues/223 is fixed
   // TODO: replace with `timers/promises` `setTimeout()` after dropping support
   // for Node <15.0.0
-  setTimeout(() => {
-    // eslint-disable-next-line unicorn/no-process-exit, n/no-process-exit
-    process.exit(EXIT_STATUS)
-  }, EXIT_TIMEOUT)
+  process.exitCode = EXIT_STATUS
+  setTimeout(forceExitProcess, EXIT_TIMEOUT).unref()
+}
+
+// Since Node 15.0.0, `unhandledRejection` makes the process exit too
+// TODO: remove after dropping support for Node <15.0.0
+const shouldExit = function (keep, reason) {
+  return (
+    !keep &&
+    (reason === 'uncaughtException' ||
+      (reason === 'unhandledRejection' &&
+        semver.gte(version, NEW_EXIT_MIN_VERSION)))
+  )
+}
+
+const NEW_EXIT_MIN_VERSION = '15.0.0'
+
+const forceExitProcess = function () {
+  // eslint-disable-next-line unicorn/no-process-exit, n/no-process-exit
+  process.exit(EXIT_STATUS)
 }
 
 const EXIT_TIMEOUT = 3000
 const EXIT_STATUS = 1
-
-export const validateExitOn = function (exitOn) {
-  if (exitOn === undefined) {
-    return
-  }
-
-  const invalidEvents = exitOn.filter((name) => !EVENTS.has(name))
-
-  if (invalidEvents.length === 0) {
-    return
-  }
-
-  throw new Error(
-    `Invalid option 'exitOn' '${invalidEvents.join(
-      ', ',
-    )}': must be one of ${EVENTS_ARR.join(', ')}`,
-  )
-}
-
-const EVENTS_ARR = [
-  'uncaughtException',
-  'unhandledRejection',
-  'rejectionHandled',
-  'warning',
-]
-const EVENTS = new Set(EVENTS_ARR)