+++ src/main.test-d.ts
@@ -5,7 +5,7 @@ import {
   expectError,
 } from 'tsd'
 
-import logProcessErrors, { Options, Level, Reason } from './main.js'
+import logProcessErrors, { Options, Reason } from './main.js'
 
 const undo = logProcessErrors()
 expectType<void>(undo())
@@ -13,11 +13,6 @@ expectError(undo(true))
 
 logProcessErrors({})
 expectAssignable<Options>({})
-
-expectAssignable<Level>('debug')
-expectNotAssignable<Level>('default')
-expectNotAssignable<Level>('silent')
-expectNotAssignable<Level>('other')
 
 expectAssignable<Reason>('warning')
 expectNotAssignable<Reason>('other')
@@ -28,8 +23,5 @@ expectError(logProcessErrors({ log: true }))
 expectError(logProcessErrors({ log(error: boolean) {} }))
 expectError(logProcessErrors({ log: () => true }))
 
-logProcessErrors({ exitOn: [] })
-logProcessErrors({ exitOn: ['info', 'debug'] })
-expectError(logProcessErrors({ exitOn: true }))
-expectError(logProcessErrors({ exitOn: [true] }))
-expectError(logProcessErrors({ exitOn: ['other'] }))
+logProcessErrors({ keep: true })
+expectError(logProcessErrors({ keep: 'true' }))