+++ src/repeat.js
@@ -34,7 +34,7 @@ const getFingerprint = function (event) {
   return fingerprintA
 }
 
-// We do not serialize `name` since this is already `name-wise`
+// We do not serialize `name` since this is already `reason`-wise
 // Key order matters since fingerprint might be truncated: we serialize short
 // and non-dynamic values first.
 const EVENT_PROPS = ['nextRejected', 'rejected', 'nextValue', 'value']